+++ src/aind_data_schema_models/coordinates.py
