+++ src/aind_data_schema_models/devices.py
